% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calculate_residual}
\alias{calculate_residual}
\title{calculate_residual}
\usage{
calculate_residual(counterfactual_frame)
}
\arguments{
\item{counterfactual_frame}{The data frame returned by read_scenario}
}
