% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{clean_all_files}
\alias{clean_all_files}
\title{clean_all_files}
\usage{
clean_all_files(filename, careful = TRUE)
}
\arguments{
\item{filename}{The same file stub argument given to setup_counterfactual}

\item{careful}{Whether to prompt before deleting files}
}
\description{
remove files generated by setup_counterfactual and run_scenario
}
