% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_scenario}
\alias{run_scenario}
\title{run_scenario}
\usage{
run_scenario(filename, init, beta_type, susceptible_type, beta_pars,
  susceptible_pars, ntime, ntrial = 1000)
}
\arguments{
\item{filename}{A string containing the desired output stub.
Should be the same argument that was passed to
\code{setup_counterfactual}.}

\item{init}{A numeric vector of the initial conditions for the
model.  It should length equal to the number of variables in the
differential equations model, and each value should be the
initial number of people in that compartment.  Should be the same
argument that was passed to \code{setup_counterfactual}.}

\item{beta_type}{One of "None" or "Flat".
"None" means that the intervention should not reduce
transmissibility.
"Flat" means that the intervention should reduce transmissibility
by a constant \code{rate} after a certain \code{start_time}.}

\item{susceptible_type}{One of "None" or "Single".
"None" means that the intervention should not eliminate
susceptibles
"Single" means that the intervention should eliminate susceptibles
at a single \code{start_time} moving \code{rate} of them from 
state \code{from_state} to state \code{to_state}.}

\item{beta_pars}{The parameters to use for the transmission
reduction function.  This should be a list.}

\item{susceptible_pars}{the parameters to use for the susceptible 
elimination function.  This should be a list.}

\item{ntime}{The number of timesteps to run the simulation over.
Should be the same argument that was passed to
\code{setup_counterfactual}}

\item{ntrial}{The number of independent simulations to run.  The 
differential equation is converted to a stochastic process when
making it agent based, and so multiple trials are recommended.
Should be the same argument that was passed to 
\code{setup_counterfactual}}
}
\value{
This function returns a data.frame one column for each
  state and one column for the simulation number.  It has one row
  for each time step, and the values are the number of people in
  that state at that time.
}
\description{
Using a previously setup counterfactual, run a
  particular scenario.
}
